requirements.txt
python-firebase>=1.2
paho-mqtt>=1.4